requirements.txt
Django==1.11.28
django_webtest>=1.9.7

gunicorn==20.0.4

dj_database_url==0.5.0
psycopg2==2.8.4

celery==4.2.1
django-picklefield==1.1.0

python-dateutil>=2.8
unicodecsv==0.14.1
bleach==3.1.1
validate_email==1.3
pycryptodome==3.8.2

python3-openid==3.0.10
boto==2.49.0
django-ses==0.8.14
oauth2client==4.1.3
rollbar==0.14.7

python-dotenv